refactor: move to vim.lsp.config() from nvim-lspconfig
