+++ lsp/buf_ls.lua
@@ -1,0 +1,5 @@
+---@type vim.lsp.Config
+return {
+    cmd = { "buf", "beta", "lsp", "--timeout=0", "--log-format=text" },
+    filetypes = { "proto" },
+}
