+++ lsp/dockerls.lua
@@ -1,0 +1,14 @@
+---@type vim.lsp.Config
+return {
+    cmd = { "docker-langserver", "--stdio" },
+    filetypes = { "dockerfile" },
+    settings = {
+        docker = {
+            languageserver = {
+                formatter = {
+                    ignoreMultilineInstructions = true,
+                },
+            },
+        },
+    },
+}
