+++ lsp/gopls.lua
@@ -1,0 +1,29 @@
+---@type vim.lsp.Config
+return {
+    cmd = { "gopls" },
+    filetypes = { "go", "gomod", "gosum" },
+    root_markers = {
+        "go.mod",
+        "go.sum",
+    },
+    settings = {
+        gopls = {
+            completeUnimported = true,
+            usePlaceholders = true,
+            gofumpt = true,
+            staticcheck = true,
+            analyses = {
+                unusedparams = true,
+            },
+            hints = {
+                assignVariableTypes = true,
+                compositeLiteralFields = true,
+                compositeLiteralTypes = true,
+                constantValues = true,
+                functionTypeParameters = true,
+                parameterNames = true,
+                rangeVariableTypes = true,
+            },
+        },
+    },
+}
