+++ lsp/lua_ls.lua
@@ -1,8 +1,14 @@
+---@type vim.lsp.Config
 return {
+  cmd = { 'lua-language-server' },
+  filetypes = { 'lua' },
+  root_markers = { '.luarc.json', '.luarc.jsonc' },
+  settings = {
     Lua = {
         hint = { enable = true },
         telemetry = { enable = false },
         diagnostics = { globals = { "vim" } },
         workspace = { checkThirdParty = false },
     },
+  },
 }
