+++ lsp/terraform_lsp.lua
@@ -1,0 +1,6 @@
+---@type vim.lsp.Config
+return {
+    cmd = { "terraform-lsp" },
+    filetypes = { "terraform", "hcl" },
+    root_markers = { ".terraform", ".git" },
+}
