+++ lua/r3x/init.lua
@@ -2,5 +2,6 @@ require("r3x.set")
 require("r3x.keymaps")
 require("r3x.lazy")
 require("r3x.terminal")
+require("r3x.lsp")
 
 vim.cmd([[colorscheme tokyonight]])
