+++ lua/r3x/lsp.lua
@@ -1,0 +1,151 @@
+---{{ LSP configuration
+vim.lsp.enable({
+    "buf_ls",
+    "dockerls",
+    "gopls",
+    "lua_ls",
+    "rust_analyzer",
+    "terraform_lsp",
+    "ts_ls",
+    "yamlls",
+})
+
+local capabilities = {
+    textDocument = {
+        foldingRange = {
+            lineFoldingOnly = true,
+            dynamicRegistration = false,
+        },
+    },
+}
+
+capabilities = require("blink.cmp").get_lsp_capabilities(capabilities)
+
+vim.lsp.config("*", { capabilities = capabilities })
+---]]
+
+vim.cmd([[autocmd FileType * set formatoptions-=ro]])
+
+vim.api.nvim_create_autocmd("TextYankPost", {
+    callback = function()
+        vim.hl.on_yank({
+            higroup = "IncSearch", -- see `:highlight` for more options
+            timeout = 200,
+        })
+    end,
+})
+
+vim.diagnostic.config({
+    virtual_text = true,
+    update_in_insert = true,
+    underline = true,
+    severity_sort = true,
+    float = {
+        focusable = true,
+        style = "minimal",
+        border = "rounded",
+        source = true,
+        header = "",
+        prefix = "",
+    },
+    signs = {
+        text = {
+            [vim.diagnostic.severity.HINT] = " ",
+            [vim.diagnostic.severity.INFO] = " ",
+            [vim.diagnostic.severity.WARN] = " ",
+            [vim.diagnostic.severity.ERROR] = " ",
+        },
+    },
+})
+
+vim.api.nvim_create_autocmd("LspAttach", {
+    desc = "LSP actions",
+    callback = function(args)
+        local client = assert(vim.lsp.get_client_by_id(args.data.client_id))
+        if not client then
+            return
+        end
+
+        ---[[ Format and autoimport on Save
+        vim.api.nvim_create_autocmd("BufWritePre", {
+            buffer = args.buf,
+            callback = function()
+                if client:supports_method("textDocument/formatting") then
+                    vim.lsp.buf.format({ bufnr = args.buf, id = client.id })
+                end
+
+                if client:supports_method("textDocument/codeAction") then
+                    local function apply_code_action(only)
+                        local actions = vim.lsp.buf.code_action({
+                            ---@diagnostic disable-next-line
+                            context = { only = only },
+                            apply = true,
+                            return_actions = true,
+                        })
+                        -- only apply if code action is available
+                        if actions and #actions > 0 then
+                            ---@diagnostic disable-next-line
+                            vim.lsp.buf.code_action({ context = { only = only }, apply = true })
+                        end
+                    end
+                    apply_code_action({ "source.fixAll" })
+                    apply_code_action({ "source.organizeImports" })
+                end
+            end,
+        })
+        ---]]
+
+        ---[[ Disable default formatting
+        if client.name == "tsserver" then
+            client.server_capabilities.documentFormattingProvider = false
+        end
+
+        if client.name == "lua_ls" then
+            client.server_capabilities.documentFormattingProvider = false
+        end
+        ---]]
+
+        ---[[ Lsp Keymaps
+        local nmap = function(keys, func, desc)
+            if desc then
+                desc = "LSP: " .. desc
+            end
+            vim.keymap.set("n", keys, func, { buffer = args.buf, noremap = true, silent = true, desc = desc })
+        end
+
+        nmap("K", vim.lsp.buf.hover, "Open hover")
+        nmap("<leader>r", vim.lsp.buf.rename, "Rename")
+        nmap("<leader>dr", vim.lsp.buf.references, "References")
+        nmap("<leader>ca", vim.lsp.buf.code_action, "Code action")
+        nmap("<leader>df", vim.lsp.buf.definition, "Goto definition")
+        nmap("<leader>ds", "<cmd>vs | lua vim.lsp.buf.definition()<cr>", "Goto definition (v-split)")
+        nmap("<leader>dh", "<cmd>sp | lua vim.lsp.buf.definition()<cr>", "Goto definition (h-split)")
+
+        -- Diagnostic
+        nmap("dn", vim.diagnostic.goto_next, "Goto next diagnostic")
+        nmap("dN", vim.diagnostic.goto_prev, "Goto prev diagnostic")
+        nmap("<leader>q", vim.diagnostic.setloclist, "Open diagnostics list")
+        nmap("<leader>e", vim.diagnostic.open_float, "Open diagnostic float")
+
+        vim.keymap.set("i", "<M-t>", vim.lsp.buf.signature_help, { buffer = args.buf })
+
+        -- inlay hints
+        nmap("<leader>lh", function()
+            vim.lsp.inlay_hint.enable(not vim.lsp.inlay_hint.is_enabled())
+        end, "Toggle inlay hints")
+
+        vim.api.nvim_buf_create_user_command(args.buf, "Fmt", function(_)
+            vim.lsp.buf.format()
+        end, { desc = "Format current buffer with LSP" })
+        ---]]
+
+        require("illuminate").configure({
+            delay = 200,
+            large_file_cutoff = 2000,
+            large_file_overrides = {
+                providers = { "lsp" },
+            },
+        })
+        require("illuminate").on_attach(client)
+    end,
+})
