--- lua/r3x/plugins/lsp/handlers.lua
@@ -1,70 +1,0 @@
-vim.diagnostic.config({
-    virtual_text = true,
-    update_in_insert = true,
-    underline = true,
-    severity_sort = true,
-    float = {
-        focusable = true,
-        style = "minimal",
-        border = "rounded",
-        source = true,
-        header = "",
-        prefix = "",
-    },
-    signs = {
-        text = {
-            [vim.diagnostic.severity.HINT] = " ",
-            [vim.diagnostic.severity.INFO] = " ",
-            [vim.diagnostic.severity.WARN] = " ",
-            [vim.diagnostic.severity.ERROR] = " ",
-        },
-    },
-})
-
-vim.api.nvim_create_autocmd("TextYankPost", {
-    callback = function()
-        vim.hl.on_yank({
-            higroup = "IncSearch", -- see `:highlight` for more options
-            timeout = 200,
-        })
-    end,
-})
-
--- format on save
-vim.api.nvim_create_autocmd("LspAttach", {
-    callback = function(args)
-        local client = vim.lsp.get_client_by_id(args.data.client_id)
-        if not client then
-            return
-        end
-
-        vim.api.nvim_create_autocmd("BufWritePre", {
-            buffer = args.buf,
-            callback = function()
-                if client:supports_method("textDocument/formatting") then
-                    vim.lsp.buf.format({ bufnr = args.buf, id = client.id })
-                end
-
-                if client:supports_method("textDocument/codeAction") then
-                    local function apply_code_action(only)
-                        local actions = vim.lsp.buf.code_action({
-                            ---@diagnostic disable-next-line
-                            context = { only = only },
-                            apply = true,
-                            return_actions = true,
-                        })
-                        -- only apply if code action is available
-                        if actions and #actions > 0 then
-                            ---@diagnostic disable-next-line
-                            vim.lsp.buf.code_action({ context = { only = only }, apply = true })
-                        end
-                    end
-                    apply_code_action({ "source.fixAll" })
-                    apply_code_action({ "source.organizeImports" })
-                end
-            end,
-        })
-    end,
-})
-
-vim.cmd([[autocmd FileType * set formatoptions-=ro]])
