--- lua/r3x/plugins/lsp/lang/go.lua
@@ -1,20 +1,0 @@
-return {
-    gopls = {
-        completeUnimported = true,
-        usePlaceholders = true,
-        gofumpt = true,
-        staticcheck = true,
-        analyses = {
-            unusedparams = true,
-        },
-        hints = {
-            assignVariableTypes = true,
-            compositeLiteralFields = true,
-            compositeLiteralTypes = true,
-            constantValues = true,
-            functionTypeParameters = true,
-            parameterNames = true,
-            rangeVariableTypes = true,
-        },
-    },
-}
