--- lua/r3x/plugins/lsp/lang/init.lua
@@ -1,9 +1,0 @@
-local M = {}
-
-M.ts = require("r3x.plugins.lsp.lang.ts")
-M.go = require("r3x.plugins.lsp.lang.go")
-M.lua = require("r3x.plugins.lsp.lang.lua")
-M.rust = require("r3x.plugins.lsp.lang.rust")
-M.yaml = require("r3x.plugins.lsp.lang.yaml")
-
-return M
