--- lua/r3x/plugins/lsp/lang/rust.lua
@@ -1,11 +1,0 @@
-return {
-    ["rust-analyzer"] = {
-        procMacro = { enable = true },
-        cargo = { allFeatures = true },
-        checkOnSave = true,
-        check = {
-            command = "clippy", --TODO: `rustup component add clippy` to install clippy
-            extraArgs = { "--no-deps" },
-        },
-    },
-}
