--- lua/r3x/plugins/lsp/lang/yamllint.lua
@@ -1,18 +1,0 @@
-return {
-    "-d",
-    [[
-        {
-            extends: relaxed,
-            rules: {
-                line-length: disable,
-                comments-indentation: disable,
-                indentation: {
-                    spaces: 2
-                }
-            }
-        }
-    ]],
-    "--format",
-    "parsable",
-    "-",
-}
